+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/geminiGoogle/GeminiGoogleClientService.kt
@@ -10,7 +10,9 @@ import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.util.text.nullize
 import dev.langchain4j.model.chat.ChatLanguageModel
+import dev.langchain4j.model.chat.StreamingChatLanguageModel
 import dev.langchain4j.model.googleai.GoogleAiGeminiChatModel
+import dev.langchain4j.model.googleai.GoogleAiGeminiStreamingChatModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -34,7 +36,16 @@ class GeminiGoogleClientService(private val cs: CoroutineScope) : LLMClientServi
             .build()
     }
 
-    override suspend fun buildStreamingChatModel(client: GeminiGoogleClientConfiguration) = null
+    override suspend fun buildStreamingChatModel(client: GeminiGoogleClientConfiguration) : StreamingChatLanguageModel {
+        val token = client.token.nullize(true) ?: retrieveToken(client.id)?.toString(true)
+        return GoogleAiGeminiStreamingChatModel.builder()
+            .apiKey(token)
+            .modelName(client.modelId)
+            .temperature(client.temperature.toDouble())
+            .topK(client.topK)
+            .topP(client.topP)
+            .build()
+    }
 
     fun saveToken(client: GeminiGoogleClientConfiguration, token: String) {
         cs.launch(Dispatchers.Default) {